Move FileEncryptor to crypto subdirectory
